# openapi/generated.openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "p4pa-classification",
    "description" : "Api and Models",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "BearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/organization/{organizationId}/classifications/{classificationId}" : {
      "get" : {
        "tags" : [ "Classifications" ],
        "summary" : "Retrieve classification detail",
        "operationId" : "getClassificationDetail",
        "parameters" : [ {
          "name" : "organizationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "classificationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassificationDetailViewDTO"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found - Classification Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassificationDetailViewDTO"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassificationDetailViewDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/organization/{organizationId}/classifications/treasured" : {
      "get" : {
        "tags" : [ "classifications", "Classifications" ],
        "description" : "The API returns a list of classifications.",
        "operationId" : "getTreasuredClassifications",
        "parameters" : [ {
          "name" : "organizationId",
          "in" : "path",
          "description" : "The ID of the organization",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "debtPositionTypeOrgCodes",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        }, {
          "name" : "label",
          "in" : "query",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/ClassificationsEnum"
          }
        }, {
          "name" : "iud",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "iuv",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "iur",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "lastClassificationDateFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "lastClassificationDateTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "payDateFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "payDateTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "paymentDateTimeFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "paymentDateTimeTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "regulationDateFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "regulationDateTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "billDateFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "billDateTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "regionValueDateFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "regionValueDateTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "pspCompanyName",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pspLastName",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "iuf",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "regulationUniqueIdentifier",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "accountRegistryCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "billAmountCents",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "remittanceInformation",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "debtorFiscalCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "billYear",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "billCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "documentYear",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "documentCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "provisionalAe",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "provisionalCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Zero-based page index (0..N)",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The size of the page to be returned",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedTreasuredClassification"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassificationErrorDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/organization/{organizationId}/classifications/paid-installments" : {
      "get" : {
        "tags" : [ "Classifications" ],
        "operationId" : "getPaidInstallments",
        "parameters" : [ {
          "name" : "organizationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "debtPositionTypeOrgCode",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "iuv",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "paymentDateTimeFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "paymentDateTimeTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "updateDateFrom",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "updateDateTo",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "iuds",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Zero-based page index (0..N)",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The size of the page to be returned",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of paid installments",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedClassificationPaidInstallmentsView"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClassificationErrorDTO"
                }
              }
            }
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ClassificationsEnum" : {
        "type" : "string",
        "enum" : [ "DOPPI", "RT_NO_IUF", "RT_NO_IUD", "IUV_NO_RT", "TES_NO_IUF_OR_IUV", "IUF_NO_TES", "IUD_RT_IUF", "RT_IUF", "RT_TES", "IUD_RT_IUF_TES", "RT_IUF_TES", "IUF_TES_DIV_IMP", "IUD_NO_RT", "TES_NO_MATCH", "UNKNOWN" ]
      },
...
      "PersonDTO" : {
        "type" : "object",
        "properties" : {
          "entityType" : {
            "$ref" : "#/components/schemas/PersonEntityType"
          },
          "fiscalCode" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "civic" : {
            "type" : "string"
          },
          "postalCode" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "nation" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "entityType", "fiscalCode", "fullName" ]
      },
      "PersonEntityType" : {
        "type" : "string",
        "enum" : [ "F", "G" ]
      },
      "ClassificationErrorDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "$ref" : "#/components/schemas/CodeEnum"
          },
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ]
      },
      "CodeEnum" : {
        "type" : "string",
        "enum" : [ "CLASSIFICATION_NOT_FOUND", "CLASSIFICATION_BAD_REQUEST", "CLASSIFICATION_CONFLICT", "CLASSIFICATION_GENERIC_ERROR" ]
      },
...
      "ClassificationDetailViewDTO" : {
        "type" : "object",
        "properties" : {
          "classificationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transferId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paymentNotificationId" : {
            "type" : "string"
          },
          "paymentsReportingId" : {
            "type" : "string"
          },
          "treasuryId" : {
            "type" : "string"
          },
          "iuf" : {
            "type" : "string"
          },
          "iud" : {
            "type" : "string"
          },
          "iuv" : {
            "type" : "string"
          },
          "iur" : {
            "type" : "string"
          },
          "transferIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "label" : {
            "$ref" : "#/components/schemas/ClassificationsEnum"
          },
          "lastClassificationDate" : {
            "type" : "string",
            "format" : "date"
          },
          "payDate" : {
            "type" : "string",
            "format" : "date"
          },
          "paymentDateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "regulationDate" : {
            "type" : "string",
            "format" : "date"
          },
          "billDate" : {
            "type" : "string",
            "format" : "date"
          },
          "regionValueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "regulationUniqueIdentifier" : {
            "type" : "string"
          },
          "accountRegistryCode" : {
            "type" : "string"
          },
          "billAmountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "remittanceInformation" : {
            "type" : "string"
          },
          "pspCompanyName" : {
            "type" : "string"
          },
          "pspLastName" : {
            "type" : "string"
          },
          "debtPositionTypeOrgCode" : {
            "type" : "string"
          },
          "installmentIngestionFlowFileName" : {
            "type" : "string"
          },
          "receiptOrgFiscalCode" : {
            "type" : "string"
          },
          "receiptPaymentReceiptId" : {
            "type" : "string"
          },
          "receiptPaymentDateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "receiptPaymentRequestId" : {
            "type" : "string"
          },
          "receiptIdPsp" : {
            "type" : "string"
          },
          "receiptPspCompanyName" : {
            "type" : "string"
          },
          "organizationEntityType" : {
            "type" : "string"
          },
          "organizationName" : {
            "type" : "string"
          },
          "receiptPayer" : {
            "$ref" : "#/components/schemas/PersonDTO"
          },
          "receiptDebtor" : {
            "$ref" : "#/components/schemas/PersonDTO"
          },
          "receiptPaymentOutcomeCode" : {
            "type" : "string"
          },
          "receiptPaymentAmount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "receiptCreditorReferenceId" : {
            "type" : "string"
          },
          "transferAmount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "transferCategory" : {
            "type" : "string"
          },
          "receiptCreationDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "installmentBalance" : {
            "type" : "string"
          },
          "billYear" : {
            "type" : "string"
          },
          "billCode" : {
            "type" : "string"
          },
          "ingestionFlowFileId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "accountCode" : {
            "type" : "string"
          },
          "domainIdCode" : {
            "type" : "string"
          },
          "transactionTypeCode" : {
            "type" : "string"
          },
          "remittanceCode" : {
            "type" : "string"
          },
          "remittanceDescription" : {
            "type" : "string"
          },
          "receptionDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "documentYear" : {
            "type" : "string"
          },
          "documentCode" : {
            "type" : "string"
          },
          "sealCode" : {
            "type" : "string"
          },
          "pspFirstName" : {
            "type" : "string"
          },
          "pspAddress" : {
            "type" : "string"
          },
          "pspPostalCode" : {
            "type" : "string"
          },
          "pspCity" : {
            "type" : "string"
          },
          "pspFiscalCode" : {
            "type" : "string"
          },
          "pspVatNumber" : {
            "type" : "string"
          },
          "abiCode" : {
            "type" : "string"
          },
          "cabCode" : {
            "type" : "string"
          },
          "ibanCode" : {
            "type" : "string"
          },
          "provisionalAe" : {
            "type" : "string"
          },
          "provisionalCode" : {
            "type" : "string"
          },
          "accountTypeCode" : {
            "type" : "string"
          },
          "processCode" : {
            "type" : "string"
          },
          "executionPgCode" : {
            "type" : "string"
          },
          "transferPgCode" : {
            "type" : "string"
          },
          "processPgNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "actualSuspensionDate" : {
            "type" : "string",
            "format" : "date"
          },
          "managementProvisionalCode" : {
            "type" : "string"
          },
          "endToEndId" : {
            "type" : "string"
          },
          "pspIdentifier" : {
            "type" : "string"
          },
          "flowDateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "senderPspType" : {
            "type" : "string"
          },
          "senderPspCode" : {
            "type" : "string"
          },
          "senderPspName" : {
            "type" : "string"
          },
          "receiverOrganizationType" : {
            "type" : "string"
          },
          "receiverOrganizationCode" : {
            "type" : "string"
          },
          "receiverOrganizationName" : {
            "type" : "string"
          },
          "totalPayments" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalAmountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "amountPaidCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paymentOutcomeCode" : {
            "type" : "string"
          },
          "acquiringDate" : {
            "type" : "string",
            "format" : "date"
          },
          "bicCodePouringBank" : {
            "type" : "string"
          },
          "paymentExecutionDate" : {
            "type" : "string",
            "format" : "date"
          },
          "paymentType" : {
            "type" : "string"
          },
          "balance" : {
            "type" : "string"
          },
          "remittanceInformationHash" : {
            "type" : "string",
            "format" : "byte"
          },
          "debtorFiscalCodeHash" : {
            "type" : "string",
            "format" : "byte"
          },
          "paymentNotificationDebtor" : {
            "$ref" : "#/components/schemas/PersonDTO"
          },
          "paymentNotificationRemittanceInformation" : {
            "type" : "string"
          },
          "paymentNotificationIud" : {
            "type" : "string"
          },
          "paymentNotificationAmountPaidCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "paymentNotificationDebtPositionTypeOrgCode" : {
            "type" : "string"
          }
        }
      },
      "PagedTreasuredClassification" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TreasuredClassificationView"
            }
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "content", "number", "size", "totalElements", "totalPages" ]
      },
...
      "ClassificationPaidInstallmentsView" : {
        "type" : "object",
        "properties" : {
          "iud" : {
            "type" : "string"
          },
          "iuv" : {
            "type" : "string"
          },
          "paymentDateTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "receiptPaymentRequestId" : {
            "type" : "string"
          },
          "organizationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "debtPositionTypeOrgCode" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "organizationId", "updateDate" ]
      },
      "PagedClassificationPaidInstallmentsView" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ClassificationPaidInstallmentsView"
            }
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int64"
          },
          "number" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "content", "number", "size", "totalElements", "totalPages" ]
      },
...
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}