# src/main/java/it/gov/pagopa/pu/classification/controller/ClassificationController.java
package it.gov.pagopa.pu.classification.controller;

import it.gov.pagopa.pu.classification.controller.generated.ClassificationsApi;
import it.gov.pagopa.pu.classification.dto.*;
import it.gov.pagopa.pu.classification.dto.generated.PagedClassificationPaidInstallmentsView;
import it.gov.pagopa.pu.classification.dto.generated.PagedTreasuredClassification;
import it.gov.pagopa.pu.classification.enums.ClassificationsEnum;
import it.gov.pagopa.pu.classification.service.ClassificationService;
import it.gov.pagopa.pu.classification.util.DateConversionUtils;
import it.gov.pagopa.pu.p4paprocessexecutions.dto.generated.LocalDateIntervalFilter;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.domain.Pageable;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;

import java.time.LocalDate;
import java.time.OffsetDateTime;
import java.util.Set;

@RestController
@Slf4j
public class ClassificationController implements ClassificationsApi {

  private final ClassificationService classificationService;

  public ClassificationController(ClassificationService classificationService) {
    this.classificationService = classificationService;
  }

  @Override
  public ResponseEntity<PagedTreasuredClassification> getTreasuredClassifications(
    Long organizationId, Set<String> debtPositionTypeOrgCodes, ClassificationsEnum label, String iud, String iuv, String iur,
    LocalDate lastClassificationDateFrom, LocalDate lastClassificationDateTo,
    LocalDate payDateFrom, LocalDate payDateTo, OffsetDateTime paymentDateTimeFrom,
    OffsetDateTime paymentDateTimeTo, LocalDate regulationDateFrom, LocalDate regulationDateTo,
    LocalDate billDateFrom, LocalDate billDateTo, LocalDate regionValueDateFrom,
    LocalDate regionValueDateTo, String pspCompanyName, String pspLastName, String iuf,
    String regulationUniqueIdentifier, String accountRegistryCode, Long billAmountCents,
    String remittanceInformation, String debtorFiscalCode,
    String billYear, String billCode, String documentYear, String documentCode,
    String provisionalAe, String provisionalCode, Pageable pageable) {
    log.info("Retrieving Classifications having organizationId: {}",
      organizationId);

    LocalDateIntervalFilter lastClassificationDate = new LocalDateIntervalFilter(
      lastClassificationDateFrom, lastClassificationDateTo);
    LocalDateIntervalFilter payDate = new LocalDateIntervalFilter(payDateFrom, payDateTo);
    OffsetDateTimeIntervalFilter paymentDate = new OffsetDateTimeIntervalFilter(
      paymentDateTimeFrom, paymentDateTimeTo);
    LocalDateIntervalFilter regulationDate = new LocalDateIntervalFilter(
      regulationDateFrom, regulationDateTo);
    LocalDateIntervalFilter billDate = new LocalDateIntervalFilter(billDateFrom,
      billDateTo);
    LocalDateIntervalFilter regionValueDate = new LocalDateIntervalFilter(
      regionValueDateFrom, regionValueDateTo);

    TreasuredClassificationFilterDTO filter = TreasuredClassificationFilterDTO.builder()
      .label(label)
      .iud(iud)
      .iuv(iuv)
      .iur(iur)
      .lastClassificationDate(lastClassificationDate)
      .payDate(payDate)
      .paymentDateTime(paymentDate)
      .regulationDate(regulationDate)
      .billDate(billDate)
      .regionValueDate(regionValueDate)
      .pspCompanyName(pspCompanyName)
      .pspLastName(pspLastName)
      .iuf(iuf)
      .regulationUniqueIdentifier(regulationUniqueIdentifier)
      .accountRegistryCode(accountRegistryCode)
      .billAmountCents(billAmountCents)
      .remittanceInformation(remittanceInformation)
      .debtorFiscalCode(debtorFiscalCode)
      .debtPositionTypeOrgCodes(debtPositionTypeOrgCodes)
      .billYear(billYear)
      .billCode(billCode)
      .documentYear(documentYear)
      .documentCode(documentCode)
      .provisionalAe(provisionalAe)
      .provisionalCode(provisionalCode)
      .build();

    return ResponseEntity.ok(
      classificationService.getPagedTreasuredClassification(organizationId,
        filter,
        pageable));
  }

  @Override
  public ResponseEntity<ClassificationDetailViewDTO> getClassificationDetail(
    Long organizationId, Long classificationId) {
    return ResponseEntity.ok(
      classificationService.getClassificationDetailView(organizationId,
        classificationId));
  }

  @Override
  public ResponseEntity<PagedClassificationPaidInstallmentsView> getPaidInstallments(
    Long organizationId, String debtPositionTypeOrgCode, String iuv, OffsetDateTime paymentDateTimeFrom, OffsetDateTime paymentDateTimeTo, OffsetDateTime updateDateFrom, OffsetDateTime updateDateTo, Set<String> iuds, Pageable pageable) {

    log.info("User requested getPaidInstallments having organizationId {}", organizationId);

    OffsetDateTimeIntervalFilter paymentDateTimeIntervalFilter = new OffsetDateTimeIntervalFilter(paymentDateTimeFrom, paymentDateTimeTo);
    LocalDateTimeIntervalFilter updateDateTimeIntervalFilter = new LocalDateTimeIntervalFilter(
      DateConversionUtils.offsetDateTime2LocalDateTime(updateDateFrom),
      DateConversionUtils.offsetDateTime2LocalDateTime(updateDateTo));

    ClassificationPaidInstallmentsFilterDTO filter = ClassificationPaidInstallmentsFilterDTO.builder()
      .iuv(iuv)
      .paymentDateTimeIntervalFilter(paymentDateTimeIntervalFilter)
      .updateDateTimeIntervalFilter(updateDateTimeIntervalFilter)
      .debtPositionTypeOrgCode(debtPositionTypeOrgCode)
      .iuds(iuds)
      .build();

    return ResponseEntity.ok(classificationService.getPaidInstallmentsView(organizationId, filter, pageable));
  }
}